+++ apps/examples/async/main.c
@@ -270,7 +270,7 @@ static void async_ctor(struct async* me, struct am_timer* timer) {
 static void ticker_task(void* param) {
     struct am_timer* timer = param;
 
-    am_task_wait_all();
+    am_task_startup_gate_wait();
 
     const int domain = AM_TICK_DOMAIN_DEFAULT;
     const uint32_t ticks_per_ms = am_time_get_tick_from_ms(domain, 1);
@@ -295,7 +295,7 @@ static void ticker_task(void* param) {
 static void input_task(void* param) {
     (void)param;
 
-    am_task_wait_all();
+    am_task_startup_gate_wait();
 
     int ch = 0;
     uint32_t prev_ms = am_time_get_ms() - (2 * ASYNC_TWO_NEWLINES_TIMEOUT_MS);