+++ apps/examples/watchdog/main.c
@@ -164,7 +164,7 @@ static void wdt_ctor(struct wdt* me, struct am_timer* timer) {
 static void ticker_task(void* param) {
     struct am_timer* timer = param;
 
-    am_task_wait_all();
+    am_task_startup_gate_wait();
 
     const int domain = AM_TICK_DOMAIN_DEFAULT;
     const uint32_t ticks_per_ms = am_time_get_tick_from_ms(domain, 1);