+++ libs/pal/freertos/pal.c
@@ -169,10 +169,10 @@ int am_get_cpu_count(void) {
 #endif
 }
 
-void am_task_wait_all(void) {}
+void am_task_startup_gate_wait(void) {}
 
-void am_task_lock_all(void) {}
+void am_task_startup_gate_close(void) {}
 
-void am_task_unlock_all(void) {}
+void am_task_startup_gate_open(void) {}
 
 void am_task_run_all(void) { vTaskStartScheduler(); }